Deprecate improved get where the type cannot be checked
If you try to get a String but the underlying type was deserialized as a another type (f.eks boolean) the `get` method would not be able to detect the wrong type and would then return the boolean, only for a ClassCastException to be thrown when the caller tried to use it as a String. 

This has been fixed by allowing the caller to specify a type that it is expected to be returned. This extra parameter can be ignored if the called specify a non-null fallback variable.